Refactor service check and update comments
The code now includes a check for the `CheckNeedOfService` function's return value. If it returns false, the `UpdateDatabase` function will return false as well. The comments related to updating the service date and distance since the last service, and updating the car's status to under service, have been removed. The comment about updating the car's status to available has been slightly modified.